+++ src/core/repart.hpp
@@ -13,7 +13,7 @@ void print_cell_info(const std::string& prefix, const std::string& method);
 
 // Get the appropriate metric function described in string "desc".
 // Throws a std::invalid_argument exception if no such metric is available
-std::function<void(std::vector<int>&)>
+std::function<void(std::vector<double>&)>
 get_metric_func(const std::string& desc);
 
 }
